app/wateja/requirements.txt
django
djangorestframework
django-countries
python-dateutil
